:sparkles: [Feat] 드롭다운 정렬 - 기사 기능 추가 & 코드 정리
정렬 상채(트리거 텍스트 정렬)
아이템 선택 시 리스트 닫기
오타 수정
상수 타입 지정
